separated header implementations to source files
implementations of function prototypes in headers are separated to their respective source files. main.c now contains main function, and header includes only.
