+++ modif_ASAL/modif_ASAL.cbp
@@ -31,6 +31,12 @@
 		<Compiler>
 			<Add option="-Wall" />
 		</Compiler>
+		<Unit filename="double_ended_linkedlist.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="file_methods.c">
+			<Option compilerVar="CC" />
+		</Unit>
 		<Unit filename="includes/double_ended_linkedlist.h" />
 		<Unit filename="includes/file_methods.h" />
 		<Unit filename="main.c">
